requirements.txt
asgiref==3.11.1
Django==6.0.2
django-environ==0.13.0
django-extensions==4.1
django-stubs==5.2.9
django-stubs-ext==5.2.9
djangorestframework==3.16.1
djangorestframework-stubs==3.16.8
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.14
inflection==0.5.1
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.5
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
uritemplate==4.2.0
## The following requirements were added by pip freeze:
certifi==2026.2.25
charset-normalizer==3.4.4
django-filter==25.2
idna==3.11
requests==2.32.5
urllib3==2.6.3